Refactor encoding and decoding mac-commands.
This is needed because of FilterListReq, which contains a variable
amount of bytes. The previous implementation was expecting a fixed
hard-coded amount of bytes per mac-command. The new implementation uses
a Cursor that is given to each mac-command decode method, such that it
can read the amount of bytes that is needed to properly decode the
mac-command payload.

This makes it possible for the FilterListReq to first read the
FilterListParam field, to know how many remaining bytes there are to
read.